extract rich style into separate module
